backend/requirements.txt
# --- Framework & Serveur ---
fastapi
uvicorn
python-multipart

# --- Base de données ---
sqlalchemy
psycopg2-binary

# --- Sécurité & Auth ---
python-dotenv
passlib[bcrypt]
python-jose[cryptography]  # Souvent utilisé avec passlib pour les JWT
python-multipart           # INDISPENSABLE pour l'upload de fichiers (FormData)
pdfplumber
pymupdf                    # Ajouté dans ton dernier script de matching
sentence-transformers
scikit-learn
openai                     # Pour la connexion au client Ollama
PyJWT

# --- Traitement de PDF ---
pdfplumber
PyMuPDF

# --- IA & Machine Learning ---
sentence-transformers
scikit-learn
openai

# --- Tests ---
pytest
httpx
